delete function added

+++ java8/src/main/java/com/example/clothingShop/repos/GoodRepo.java
@@ -3,5 +3,5 @@ package com.example.clothingShop.repos;
 import com.example.clothingShop.domain.Good;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GoodRepo extends CrudRepository<Good, Long> {
+public interface GoodRepo extends CrudRepository<Good, Integer> {
 }
